+++ RoughSets/Infovision.Datamining.RoughSet/BireductGamma.cs
@@ -8,17 +8,17 @@ namespace Infovision.Datamining.Roughset
     {
         #region Constructors
 
-        public BireductGamma(DataStore dataStore, int[] fieldIds, int[] objectIndexes, double epsilon)
+        public BireductGamma(DataStore dataStore, int[] fieldIds, int[] objectIndexes, decimal epsilon)
             : base(dataStore, fieldIds, objectIndexes, epsilon)
         {
         }
 
-        public BireductGamma(DataStore dataStore, int[] fieldIds, double epsilon)
+        public BireductGamma(DataStore dataStore, int[] fieldIds, decimal epsilon)
             : this(dataStore, fieldIds, new int[] { }, epsilon)
         {
         }
 
-        public BireductGamma(DataStore dataStore, double epsilon)
+        public BireductGamma(DataStore dataStore, decimal epsilon)
             : this(dataStore, dataStore.DataStoreInfo.GetFieldIds(FieldTypes.Standard), new int[] { }, epsilon)
         {
         }
